Make some aspects of the project swappable by external integrations (#235)
# Purpose

This purpose of this pull request is to make it possible for advanced
integrators to swap out some of the dependencies with others.

For a first iteration, I did not want to require current consumers of
this library to have to make any changes - for them this pull requests
changes nothing whatsoever.
This means that I did not do any sort of of changes to the structure or
name of the Maven modules. I believe that if we want to do that, if can
be done at a latter point.

# Changes

## Gson

All this change does is move Gson out of the model classes. This is
needed by integrations that plan to exclude Gson from the classpath, so
that the models can continue to be loaded without incurring ClassLoading
issues.

## Template SPI

I introduced `PromptTemplateFactory` which can be used for integrations
to supply their own version of the template engine.
Implementations are loaded using Java's ServiceLoader mechanism and if
none are found, the default `MustachePromptTemplateFactory` is used.

I have leveraged this SPI my Quarkus integration
[here](https://github.com/geoand/quarkus-langchain4j/blob/487711af4c7b465dbc40b90314c67de712ab8a15/runtime/src/main/java/io/quarkiverse/langchain4j/QuarkusPromptTemplateFactory.java).

## Make OpenAiStreamingResponseBuilder thread-safe

This class needs to be thread safe because it is called when a streaming
result comes back
and there is no guarantee that this thread will be the same as the one
that initiated the request