Allow reading prompt (both system & user) message from a resource (#527)
This is to implement the suggestion done in
https://github.com/langchain4j/langchain4j/discussions/508 :

It would be valuable to our team if we could specify a resource to read
as a prompt-message. We use this technique for more things like
SQL-queries. Makes reading and crafting these texts easier.

I would be happy to add this to the source via a PR but I thought let's
test the waters first to see if more people would like this and how high
the chance would be of the PR getting pulled.

It would work like this: where the annotation is read (eg.
DefaultAiServices.prepareUserMessage) we could switch on available
parameters to the annotation and read the resource specified (based on
the methods-class using clazz.getResourceAsStream) in stead of using the
provided text. (I do realize we could want to cache the resource reading
when that becomes a bottleneck later)

What would you think? What would the names of the parameters to
@SystemMessage & @Usermessage need to be? Something like "resource" or
"resourceName"?

This is the code I could use to text resource:
```
public static String getString(Class<?> clazz, String name) {
    return getString(clazz.getResourceAsStream(name));
}

public static String getString(InputStream inputStream) {
    try (Scanner scanner = new Scanner(inputStream); //
            Scanner s = scanner.useDelimiter("\\A")) {
        return s.hasNext() ? s.next() : "";
    }
}
```

Any suggestions in this area would be welcome too.