Add advanced RAG with Azure AI Search (#587)
This PR should fix #576 and add advanced RAG with hybrid search and
semantic re-ranking with Azure AI Search.

In the current implementation, the scoring for full text search, hybrid
search and semantic search are done using comments directly from the
Azure AI Search team, as it seems the documentation is only correct for
vector search. Have a look at the `fromAzureScoreToRelevanceScore`
function for more information.