Allow to programmatically define system and user messages in the AiServices (#862)
This pull request is a variation and simplification of what I proposed
[here](https://github.com/langchain4j/langchain4j/pull/861). As
anticipated there instead of making the AiServices aware of the concept
of a state machine I simply added the possibility of making the system
and user messages configurable in a programmatic way. Currently this
pull request has the same limitation of the other one: lack of the
possibility of configuring messages for each user/conversation and of
support of tools, but I believe that both things could be added once we
agree on the main direction.