Update ModelListener API to be more integration friendly
The idea here is to allow the onRequest method to return
an object which will then be passed along to the
response and error methods.
This allows for integrations to attach anything they
need when the request is made and access it when the
response or error methods are called.

On the implementation side this looks a little
awkward because of the generics, but the upside
is that for users / integrators of ModelListener
the signature of the class and methods provide very good
hints on how to implement them