Add docusaurus-lunr-search plugin (#1221)
## Issue
[https://github.com/langchain4j/langchain4j/issues/1219](https://github.com/langchain4j/langchain4j/issues/1219)


## Change
Added the `docusaurus-lunr-search` plugin.

It may seem that for search feature we would need a server. But nothing
stops us from passing that logic to the client)

So, if the index for pages is not too big, we will delegate the search
functionality to client. That's what this plugin does (I think).

It works with `npm run serve`, but I couldn't test it with just opening
the `index.html` file.

Oh, I'm sorry, my editor autoformatted the file)

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [x] There are no breaking changes
- [ ] I have added unit and integration tests for my change
- [ ] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
- [ ] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
<!-- Before adding documentation and example(s) (below), please wait
until the PR is reviewed and approved. -->
- [ ] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)