DefaultRetrievalAugmentor: process in the same thread when single query and single retriever, otherwise use executor and cache threads for 1 second (#1479)
## Issue
This PR partially fixes #1454

## Context
`DefaultRetrievalAugmentor` currently uses an `Executor` to parallelize
the processing (consider multiple `Query`s and/or multiple
`ContentRetriever`s).
The default `Executor` instance caches (non-daemon) threads for 60
seconds, so when the application is ready to shut down, it can hang for
another 60 seconds before it can actually exit.
For the majority of the use cases (single `Query` and single
`ContentRetriever`) there is no need to use an `Executor`, processing
can be done in the same thread without an `Executor`, thus there will be
no hanging.
For the rest of the use cases we can use the `Executor` to parallelize
the processing as before. But for default `Executor`, reduce the time
from 60 to 1 second, which makes "handing time" acceptable.
In any case, the user can always provide a custom instance of an
`Executor` and manage it externally.


## Change
- Changes in `DefaultRetrievalAugmentor`:
- When there is only a single `Query` and a single `ContentRetriever`
(majority of use cases), processing is done in the same thread
(`Executor` is not used at all)
- Otherwise, the `Executor`is used to parallelize query routing and
content retrieval. The default `Executor` now caches threads for 1
second (instead of 60 seconds)
  - Added javadoc and documentation
- Added documentation for
https://github.com/langchain4j/langchain4j/issues/1454



## General checklist
- [X] There are no breaking changes
- [X] I have added unit and integration tests for my change
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
- [X] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)
- [ ] I have added/updated [Spring Boot
starter(s)](https://github.com/langchain4j/langchain4j-spring) (if
applicable)