Automatically review PRs with OpenRewrite recipe runs (#673)
As discussed on
https://github.com/langchain4j/langchain4j/pull/622#pullrequestreview-1895918197,
this PR adds a workflow to automatically review pull requests. The rules
are configurable and can be extended; I'd thought to start small, and
then expand in separate pull requests where I both make the changes to
the project, and add the recipe to the list of suggestions for future
pull requests.

As an example of what this looks like in practice, have a look at
https://github.com/openrewrite/rewrite-migrate-java/pull/421
Notice how we have a bot `github-actions` that places code suggestion
comments that can be quickly applied by the user. This should make for a
reasonably friction-free enforcement of code standards going forward,
using mostly standard components +
https://github.com/googleapis/code-suggester


![image](https://github.com/langchain4j/langchain4j/assets/1027334/b29e0357-111c-4d94-b3df-e15c672330b2)


I'll self review with a few small callouts below, just to make review
easier.

Assumes that we first merge my other pull request, such that we don't
get any suggestions for items fixed there.
- #622