Migrated the "responsible AI" integration tests to GitHub Models (#1945)
This follows the discussion at
https://github.com/langchain4j/langchain4j/pull/764

Here are my findings:

- New version of the OpenAI models are really good at filtering
violence/self harm/etc so it's very hard to hit the responsible AI
filters.
- To my surprise, other models are also quite good, including the
Mistral ones. For this tests, I'm using PHI 3.5 mini, as usually the
smallest models are less good at everything, which includes filtering.
- I used to test the JSON response with the Azure OpenAI API, which
included some information about the type of filters (violence, self
harm, etc), but I can't access this data with the Azure AI Inference
API. So now I'm testing with the "finish reason", which is less precise
but tells for sure if the content was filtered.

So I'm not super satisfied with this PR, but it works and does some
basic responsible AI tests, so we know the API works.