Health checks for MCP clients (#2769)
## Issue
<!-- Please specify the ID of the issue this PR is addressing. For
example: "Closes #1234" or "Fixes #1234" -->
Closes https://github.com/langchain4j/langchain4j/issues/2766

## Change
Adds the `McpClient.checkHealth()` method that throws an exception if
the server isn't reachable or isn't responding to pings within a
reasonable timeout. I didn't include a dependency on MicroProfile Health
as this probably isn't suitable here, but any runtime that integrates it
(Quarkus) may want to wrap an invocation of this method into a built-in
MicroProfile Health check.


## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [x] There are no breaking changes
- [x] I have added unit and/or integration tests for my change
- [x] The tests cover both positive and negative cases
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green