Fix build after apache-poi upgrade (#2849)
<!--
Thank you so much for your contribution!

Please fill in all the sections below.
Please open the PR as a draft initially. Once it is reviewed and
approved, we will ask you to add documentation and examples.
Please note that PRs with breaking changes or without tests will be
rejected.

Please note that PRs will be reviewed based on the priority of the
issues they address.
We ask for your patience. We are doing our best to review your PR as
quickly as possible.
Please refrain from pinging and asking when it will be reviewed. Thank
you for understanding!
-->

## Issue
<!-- Please specify the ID of the issue this PR is addressing. For
example: "Closes #1234" or "Fixes #1234" -->

After this upgrade
https://github.com/langchain4j/langchain4j/commit/fa6b02c50182331124112efcf7956364bb97bb5f
the build fails with this error message

```
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-enforcer-plugin:3.5.0:enforce (enforce) on project langchain4j-document-parser-apache-poi: 
[ERROR] Rule 1: org.apache.maven.enforcer.rules.dependency.DependencyConvergence failed with message:
[ERROR] Failed while enforcing releasability.
[ERROR] 
[ERROR] Dependency convergence error for commons-io:commons-io:jar:2.18.0 paths to dependency are:
[ERROR] +-dev.langchain4j:langchain4j-document-parser-apache-poi:jar:1.0.0-beta3-SNAPSHOT
[ERROR]   +-org.apache.poi:poi:jar:5.4.0:compile
[ERROR]     +-commons-io:commons-io:jar:2.18.0:compile
[ERROR] and
[ERROR] +-dev.langchain4j:langchain4j-document-parser-apache-poi:jar:1.0.0-beta3-SNAPSHOT
[ERROR]   +-org.apache.poi:poi-ooxml:jar:5.4.0:compile
[ERROR]     +-org.apache.commons:commons-compress:jar:1.27.1:compile
[ERROR]       +-commons-io:commons-io:jar:2.16.1:compile
[ERROR] and
[ERROR] +-dev.langchain4j:langchain4j-document-parser-apache-poi:jar:1.0.0-beta3-SNAPSHOT
[ERROR]   +-org.apache.poi:poi-ooxml:jar:5.4.0:compile
[ERROR]     +-commons-io:commons-io:jar:2.18.0:compile
[ERROR] and
[ERROR] +-dev.langchain4j:langchain4j-document-parser-apache-poi:jar:1.0.0-beta3-SNAPSHOT
[ERROR]   +-org.apache.poi:poi-scratchpad:jar:5.4.0:compile
[ERROR]     +-commons-io:commons-io:jar:2.18.0:compile

```

## Change
(Temporarily?) exclude `commons-io` from `poi-ooxml` so it will take the
version coming from `poi`.

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [X] There are no breaking changes
- [ ] I have added unit and/or integration tests for my change
- [ ] The tests cover both positive and negative cases
- [ ] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [ ] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
<!-- Before adding documentation and example(s) (below), please wait
until the PR is reviewed and approved. -->
- [ ] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)
- [ ] I have added/updated [Spring Boot
starter(s)](https://github.com/langchain4j/langchain4j-spring) (if
applicable)