Allowing caller to pass in class loader for classpath document loading (#3090)
<!--
Thank you so much for your contribution!

Please fill in all the sections below.
Please open the PR as a draft initially. Once it is reviewed and
approved, we will ask you to add documentation and examples.
Please note that PRs with breaking changes or without tests will be
rejected.

Please note that PRs will be reviewed based on the priority of the
issues they address.
We ask for your patience. We are doing our best to review your PR as
quickly as possible.
Please refrain from pinging and asking when it will be reviewed. Thank
you for understanding!
-->

## Change
<!-- Please describe the changes you made. -->
This is an enhancement to the classpath document loading to allow the
caller to pass in which class loader to use rather than dependending on
the default one. Quarkus in particular needs to do this.

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [ X ] There are no breaking changes
- [ X ] The tests cover both positive and negative cases
- [ X ] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [ X ] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green